# Perf: Use owned buffers