--- a/perf/src/server.rs
+++ b/perf/src/server.rs
@@ -174,12 +174,14 @@ async fn drain_stream(mut stream: quinn::RecvStream) -> Result<()> {
 
 async fn respond(mut bytes: u64, mut stream: quinn::SendStream) -> Result<()> {
     const DATA: [u8; 1024 * 1024] = [42; 1024 * 1024];
+
     while bytes > 0 {
-        let n = stream
-            .write(&DATA[..bytes.min(DATA.len() as u64) as usize])
+        let chunk_len = bytes.min(DATA.len() as u64);
+        stream
+            .write_chunk(Bytes::from_static(&DATA[..chunk_len as usize]))
             .await
             .context("sending response")?;
-        bytes -= n as u64;
+        bytes -= chunk_len;
     }
     trace!("finished responding on {}", stream.id());
     Ok(())