@Matthias247 2021-03-02
> Should we replace the `const`s with a single `Bytes` that gets cloned to avoid some more allocs/copying? I guess it's big enough now that the difference might not register.

`Bytes::from_static` has a vtable which doesn't have any copies or allocations